+++ src/XSharp.VsParser.Helpers/Parser/ParserHelper.cs
@@ -8,13 +8,11 @@ using System.Diagnostics;
 using System.IO;
 using System.Text;
 using System.Xml.Linq;
-using XSharp.VsParser.Helpers.Rewriters;
 
 namespace XSharp.VsParser.Helpers.Parser
 {
     public class ParserHelper
     {
-        const string NewList = "$NewLine";
         GenericErrorListener _ErrorListener = new();
         string _SourceCode;
         XSharpParseOptions _XSharpOptions;
@@ -25,7 +23,7 @@ namespace XSharp.VsParser.Helpers.Parser
 
         public AbstractSyntaxTree SourceTree { get; }
 
-        void CheckParseSuccessful()
+        internal void CheckParseSuccessful()
         {
             if (_StartRule == null)
                 throw new ArgumentException("Parsing was not successful");
@@ -83,122 +81,6 @@ namespace XSharp.VsParser.Helpers.Parser
 
             foreach (var listener in listeners)
                 ParseTreeWalker.Default.Walk(listener, _StartRule);
-        }
-
-        public (bool changed, string newSourceCode) ExecuteRewriters(List<XSharpBaseRewriter> rewriters)
-        {
-            Debug.Assert((rewriters?.Count ?? 0) > 0, "List of rewriters can not be empty");
-            CheckParseSuccessful();
-
-            var tokenRewriter = new TokenStreamRewriter(_Tokens);
-
-            foreach (var rewriter in rewriters)
-            {
-                rewriter.Initialize(tokenRewriter);
-                ParseTreeWalker.Default.Walk(rewriter, _StartRule);
-            }
-
-            var newSourceCode = tokenRewriter.GetText();
-            return (_SourceCode != newSourceCode, newSourceCode);
-        }
-
-        public bool ExecuteRewriters(List<XSharpBaseRewriter> rewriters, string newFilename = null)
-        {
-            var (changed, newSourceCode) = ExecuteRewriters(rewriters);
-            if (changed)
-            {
-                newFilename ??= _FileName;
-                File.WriteAllText(newFilename, newSourceCode);
-            }
-
-            return changed;
-        }
-
-
-        /// <summary>
-        /// Dumps the AST created by parsing as XDocument
-        /// </summary>
-        public XDocument DumpXml()
-        {
-            CheckParseSuccessful();
-
-            static void DumpTerminalValue(TerminalNodeImpl terminalNodeImpl, XElement element)
-            {
-                var text = terminalNodeImpl.Payload.Text;
-                if (text == Environment.NewLine)
-                    element.Add(NewList);
-                else if (text.IndexOfAny(new char[] { '\n', '<', '>', '&' }) == -1)
-                    element.Add(text);
-                else
-                    element.Add(new XCData(text));
-            }
-
-            static void DumpElement(IParseTree rule, XContainer container)
-            {
-                var node = new XElement(rule.GetType().Name);
-                container.Add(node);
-
-                if (rule is TerminalNodeImpl terminalNodeImpl && rule.ChildCount == 0)
-                    DumpTerminalValue(terminalNodeImpl, node);
-
-                for (int i = 0; i < rule.ChildCount; i++)
-                    DumpElement(rule.GetChild(i), node);
-            }
-
-            var doc = new XDocument();
-            DumpElement(_StartRule, doc);
-            return doc;
-        }
-
-
-        /// <summary>
-        /// Dumps the AST created by parsing as Yaml
-        /// </summary>
-        public string DumpYaml()
-        {
-            CheckParseSuccessful();
-            var sb = new StringBuilder();
-
-            void DumpTerminalValue(string text, int indent)
-            {
-                if (text == Environment.NewLine)
-                {
-                    sb.AppendLine(" " + NewList);
-                    return;
-                }
-
-                text = text.Replace("\\", "\\\\").Replace("\"", "\\\"");
-                if (text.Contains("\n"))
-                {
-                    sb.AppendLine(" |-");
-                    foreach (var line in text.Replace("\r", "").Split('\n'))
-                        sb.AppendLine(new string(' ', indent * 2) + line);
-                }
-                else
-                    sb.Append(" \"").Append(text).AppendLine("\"");
-            }
-
-            void DumpElement(IParseTree rule, int indent)
-            {
-                var indentString = new string(' ', indent * 2);
-                indent++;
-
-                sb.Append(indentString).Append("- ").Append(rule.GetType().Name).Append(":");
-
-                if (rule is TerminalNodeImpl terminalNodeImpl && rule.ChildCount == 0)
-                    DumpTerminalValue(terminalNodeImpl.Payload.Text, indent + 1);
-                else
-                {
-                    sb.AppendLine();
-                    for (int i = 0; i < rule.ChildCount; i++)
-                        DumpElement(rule.GetChild(i), indent);
-                }
-            }
-
-            sb.Append(_StartRule.GetType().Name).AppendLine(":");
-            for (int i = 0; i < _StartRule.ChildCount; i++)
-                DumpElement(_StartRule.GetChild(i), 0);
-            return sb.ToString();
         }
 
         #region Builders
